Refactored Sam into component
